Update GreenvilleRevenue.cs
Added a constant for the entry fee and used it to calculate this year's expected revenue. Added a bool and if/else to display whether or not this year has more contestants. I got a bit carried away and also calculated the difference between each year's turnout and a statement for if they are both equal (because I just really enjoy programming).